pyproject.toml
[project]
name = "iam57-onanibot"
version = "0.1.0"
description = "iam57-onanibot"
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = ["nonebot-adapter-onebot>=2.4.6", "nonebot2[fastapi]>=2.4.2", "jmcomic>=2.5.32", "pillow>=11.1.0", "pyyaml>=6.0.2", "requests>=2.32.3"]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
plugin_dirs = ["iam57_onanibot/plugins"]
builtin_plugins = []
